% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversion.R
\name{a_freq_list}
\alias{a_freq_list}
\title{Convert data frame of allele frequencies to nested lists}
\usage{
a_freq_list(D, pop_level = "collection")
}
\arguments{
\item{D}{the long-format dataframe of counts by collection, locus, and allele,
output by \code{reference_allele_counts}, to be made into a nested list}
}
\value{
\code{a_freq_list} returns a list named by loci, each element of which is a matrix
containing that locus's allele count data. Rows in the matrix mark alleles, and columns collections
}
\description{
List-izes the output of \code{reference_allele_counts} into a usable format for \code{allelic_list}
}
\examples{

 # Generate a list of individual genotypes by allele from
 # the alewife data's reference allele count tables
 example(reference_allele_counts)
 ale_ac <- a_freq_list(ale_rac)
}
\keyword{internal}
