% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversion.R
\name{avg_assign_coll}
\alias{avg_assign_coll}
\title{generate a matrix of the average (by collection) scaled likelihood of fish assigned to a reporting unit}
\usage{
avg_assign_coll(SL, coll, RU_starts, RU_vec)
}
\arguments{
\item{SL}{a scaled likelihood matrix}

\item{coll}{a vector of the collection of origin indices of the fish (length == ncol(SL))}

\item{RU_starts}{a vector delineating starting indices of different reporting units in RU_vec}

\item{RU_vec}{a vector of collection indices, organized by reporting unit}
}
\description{
generate a matrix of the average (by collection) scaled likelihood of fish assigned to a reporting unit
}
\examples{
params <- tcf2param_list(alewife, 15)
SL <- geno_logL(params) \%>\% exp() \%>\% apply(2, function(x) x/sum(x))
avg_SL_coll2ru <- avg_assign_coll(SL, params$coll, params$RU_starts, params$RU_vec)

}

