% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversion.R
\name{avg_miss_ru}
\alias{avg_miss_ru}
\title{generate a matrix of the average percentage of a reporting unit misassigned}
\usage{
avg_miss_ru(SL, coll, RU_starts, RU_vec)
}
\arguments{
\item{SL}{a scaled likelihood matrix}

\item{params}{the MCMC parameter list output by list_diploid_params and tcf2param_list}
}
\description{
generate a matrix of the average percentage of a reporting unit misassigned
}
\examples{
params <- tcf2param_list(alewife, 15)
SL <- geno_logL(params) \%>\% exp() \%>\% apply(2, function(x) x/sum(x))
miss <- avg_miss_ru(SL, params$coll, params$RU_starts, params$RU_vec)

}

