% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversion.R
\name{avg_SL_coll}
\alias{avg_SL_coll}
\title{generate a matrix of the average percentage of a collection assigned to each collection}
\usage{
avg_SL_coll(SL, coll)
}
\arguments{
\item{SL}{a scaled likelihood matrix}

\item{coll}{a vector of the collections of origin of the columns in SL}
}
\value{
a matrix in which each column represents the collection of origin, and each row
the average scaled likelihood of assignment to that collection
}
\description{
generate a matrix of the average percentage of a collection assigned to each collection
}
\examples{
params <- tcf2param_list(alewife, 15)
SL <- geno_logL(params) \%>\% exp() \%>\% apply(2, function(x) x/sum(x))
avg_SL <- avg_SL_coll(SL, params$coll)

}

