% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_correction.R
\name{bias_comparison}
\alias{bias_comparison}
\title{Test the effects of bias correction on a particular reference dataset}
\usage{
bias_comparison(reference, gen_start_col, seed = 5)
}
\arguments{
\item{reference}{a two-column format genetic dataset, with a "repunit" column
specifying each individual's reporting unit of origin, a "collection" column
specifying the collection (population or time of sampling) and "indiv" providing
a unique name}

\item{gen_start_col}{the first column containing genetic data in \code{reference}.
All columns should be genetic format following this column, and gene copies from the
same locus should be adjacent}

\item{seed}{the random seed for simulations}
}
\value{
\code{bias_comparison} returns a list; the first element is
a list of the relevant rho values generated on each iteration of the random "mixture"
creation. This includes the true rho value, the standard result \code{rho_mcmc},
the misassignment-scaled \code{rho_bh}, and the parametric bootstrapped \code{rho_pb}.
The second element is a dataframe listing summary statistics for each
reporting unit and estimation method
}
\description{
Takes a reference two-column genetic dataset, pulls a series of random
"mixture" datasets with varying reporting unit proportions from this reference,
and compares the results of GSI through standard MCMC, misassignment-scaled MCMC,
and parametric-bootstrap MCMC bias correction
}
\examples{
ale_bias <- bias_comparison(alewife, 15)

}

