% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blueback}
\alias{blueback}
\title{Microsat data from blueback herring reference populations}
\format{
A tibble.
}
\source{
\url{https://datadryad.org/stash/dataset/doi:10.5061/dryad.80f4f}
}
\description{
Standard two-column genetic data with lots of other columns preceding it.
Can be fed directly into rubias because it has at least the columns sample_type,
collection, repunit and indiv.
}
