% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_pipeline.R
\name{C_test}
\alias{C_test}
\title{Generate a matrix of average scaled likelihoods from a mixture dataset}
\usage{
C_test(D, gen_start_col)
}
\arguments{
\item{D}{a two-column genetic dataset with "repunit", "collection", and "sample_type" columns}

\item{gen_start_col}{the first column of genetic data}
}
\description{
Generate a matrix of average scaled likelihoods from a mixture dataset
}
\examples{
D <- rbind(alewife, alewife)
D$repunit <- as.character(D$repunit)
D$collection <- as.character(D$collection)
D$sample_type[1071:2140] <- "mixture"
D$repunit[1071:2140] <- "mix"
D$collection[1071:2140] <- "mix"
D$repunit <- factor(D$repunit)
D$collection <- factor(D$collection)
gen_start_col <- 15

}

