% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_known_collections.R
\name{check_known_collections}
\alias{check_known_collections}
\title{check a baseline and mixture file together to ensure the known_collections are valid if they exist}
\usage{
check_known_collections(R, M)
}
\arguments{
\item{R}{reference data frame}

\item{M}{mixture data frame}
}
\description{
Simple function that checks for known_collections columns in a reference and mixture and
makes sure that they are compliant.
If there is a non-NA entry in the Mixture frame's known_collection column this
function returns TRUE.  Otherwise it returns FALSE.
}
\keyword{internal}
