% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eca-internal-funcs.R
\name{check_refmix}
\alias{check_refmix}
\title{A helper function to check that the input data frame is OK}
\usage{
check_refmix(D, gen_start_col, type = "reference")
}
\arguments{
\item{D}{the data frame}

\item{gen_start_col}{the column in which the genetic data starts}

\item{type}{For writing errors, supply "mixture" or "reference" as appropriate.}
}
\description{
Just checks to make sure that column types are correct.
}
\details{
It also checks the patterns of missing data, and from that infers whether
markers are haploid or diploid.
}
\keyword{internal}
