% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chinook_all}
\alias{chinook_all}
\title{SNP data from Chinook reference and mixtures mixed together}
\source{
Southwest Fisheries Science Center, Santa Cruz, CA
}
\description{
gotta say more
}
