% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chinook_mix}
\alias{chinook_mix}
\title{SNP data from Chinook salmon taken in May/August 2015 from California fisheries}
\format{
A tbl_df-ed (from dplyr) data frame with 2256 rows and 193 variables. The first
four columns are meta data.
The remaining columns are two columns for each locus.  These columns are named like,
"Locus.1" and "Locus.2" for the first and second gene copies at that locus.  For example,
"Ots_104569-86.1" and  "Ots_104569-86.2".  The locus columns are ints and missing data
is denoted by NA.
}
\source{
Southwest Fisheries Science Center, Santa Cruz, CA
}
\description{
This has data from 91 SNP markers (a subset of the 95 markers in the \code{\link{chinook}} baseline
data set).
}
