% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chinook_repunit_levels}
\alias{chinook_repunit_levels}
\title{a vector that gives a desired sort order of the chinook repunits}
\source{
Made it up!
}
\description{
This is just an example of what one would use as levels in order to
get the \code{\link{chinook}} repunits in a desired sort order after
analysis.  The issue here is that repunit in the input data frame to most
functions must be a character vector, not a factor.  But, after analysis
you can always make them a factor again and use a vector like this
one to specify the levels.
}
