% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coll_ac}
\alias{coll_ac}
\title{Draw all allele counts for a population from a vector of allele counts}
\usage{
coll_ac(l, c, ac_v, L, C, A, CA)
}
\arguments{
\item{l}{the locus desired}

\item{c}{the collection index desired}

\item{ac_v}{the vector of allele counts}

\item{L}{the total number of loci}

\item{C}{the total number of collections}

\item{A}{an integer vector of alleles in each population}

\item{CA}{an integer vector of the cumulative number of alleles at all previous loci in the vector}
}
\description{
(This is a test for our allele count vector indexing macro)
}

