% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mixture_z_scores.R
\name{compute_mixture_z_scores}
\alias{compute_mixture_z_scores}
\title{compute empirical Z-scores for mixture sample population individuals}
\usage{
compute_mixture_z_scores(SA)
}
\arguments{
\item{SA}{The tibble that is returned by \code{\link{self_assign}} using the same reference.}

\item{M}{The tibble that is output by \code{\link{infer_mixture}}.}
}
\value{
A tibble like with a few more columns
}
\description{
To run this, you must supply the z-scores and logls of the baseline fish, too.
}
