% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_missing_data.R
\name{count_missing_data}
\alias{count_missing_data}
\title{for each individual count the number or loci missing and non_missing}
\usage{
count_missing_data(D, gen_start_col)
}
\arguments{
\item{D}{the data frame}

\item{gen_start_col}{the column in which the genetic data starts}
}
\value{
returns a data frame with indiv (as characters), n_non_miss_loci, n_miss_loci (as numeric) and missing_loci
(as a list-column of named integer vectors)
}
\description{
Takes a rubias genetic data frame that must have column "indiv". Haploids have the
second column at each locus totally missing.  Diploids with missing data will have both gene copies
missing.
}
\keyword{internal}
