% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dirch_from_counts}
\alias{dirch_from_counts}
\title{given a vector of different categories in 1...n and a prior simulate a dirichlet r.v.}
\usage{
dirch_from_counts(C, lambda)
}
\arguments{
\item{C}{a vector giving counts of categories}

\item{pi}{priors for the categories}
}
\description{
The categories are labeled in C from 1 up to n.  n is the length of pi which is vector of priors.
Note that all elements of pi must be strictly greater than 0.
}

