% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversion.R
\name{DP_sum}
\alias{DP_sum}
\title{Get summed Dirichlet parameters for all alleles and populations in a numeric vector}
\usage{
DP_sum(DP_list)
}
\arguments{
\item{DP}{the list of dirichlet parameter matrices}
}
\description{
Get summed Dirichlet parameters for all alleles and populations in a numeric vector
}
\examples{
example(allelic_list)
as <- lapply(ale_ac, nrow)
loc_cycle <- names(ale_ac)
names(loc_cycle) <- loc_cycle
ale_DP_list <- lapply(loc_cycle, function(x) ale_ac[[x]] + 1/as[[x]])
sum_DP <- DP_sum(ale_DP_list)

}

