% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{geno_logL_RU}
\alias{geno_logL_RU}
\title{Calculate list of genotype likelihoods}
\usage{
geno_logL_RU(par_list)
}
\arguments{
\item{par_list}{genetic data converted to the param_list format by tcf2param_list}
}
\description{
Calculate list of genotype likelihoods
}
\examples{
example(tcf2param_list)
ale_glL <- geno_logL(ale_par_list)
}

