% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{genotype_i}
\alias{genotype_i}
\title{Fetch individual genotypes from a locus and population}
\usage{
genotype_i(l, i, I_v, P, I)
}
\arguments{
\item{l}{the index of the desired locus}

\item{i}{the index of the desired individual to sample}

\item{I_v}{a vector of individual genotypes, created by unlisting allelic_list output}

\item{P}{Ploidy}

\item{I}{Total number of individuals}
}
\description{
(This is a test case for our individual indexing macro)
}

