% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eca-internal-funcs.R
\name{get_ploidy_from_frame}
\alias{get_ploidy_from_frame}
\title{infer the ploidy from the pattern of NAs in the columns of data}
\usage{
get_ploidy_from_frame(tmp, type)
}
\arguments{
\item{tmp}{a data frame with 2 * L columns (two for each locus)}
}
\description{
This is strictly internal
}
\keyword{internal}
