% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repuclust.R
\name{merge_repuclusts}
\alias{merge_repuclusts}
\title{simple function to merge the repuclusts in the pair_posts data frame}
\usage{
merge_repuclusts(pp, new_repu, d1, d2)
}
\arguments{
\item{pp}{the pair_posts data frame}

\item{new_repu}{the index of the new reporting unit node}

\item{d1}{the repuclust index of the first daughter of new_repu}

\item{d2}{the repuclust index of the second daughter of new_repu}
}
\description{
simple function to merge the repuclusts in the pair_posts data frame
}
\keyword{internal}
