% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_pipeline.R
\name{mixture_simulation_pipeline}
\alias{mixture_simulation_pipeline}
\title{Generate population proportion estimates from a reference and desired collection composition}
\usage{
mixture_simulation_pipeline(reference, gen_start_col)
}
\arguments{
\item{reference}{a two-column format genetic dataset, with "repunit", "collection", and "indiv"
columns}

\item{gen_start_col}{the first column of genetic data in reference}
}
\description{
Using a reference allele count matrix and a vector specifying from which collection to draw
each individual, creates a genotype-logL matrix based on simulation-by-individual,
then uses this in three different estimates of population mixture proportions: EM, MCMC,
and hierarchical MCMC
}

