% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/modify-scaled-likelihoods-for-known-mixture-fish.R
\name{modify_scaled_likelihoods_for_known_mixture_fish}
\alias{modify_scaled_likelihoods_for_known_mixture_fish}
\title{for individuals of known origin in the mixture, put all their weight on their known collection}
\usage{
modify_scaled_likelihoods_for_known_mixture_fish(SL, KC, CFL)
}
\arguments{
\item{SL}{the matrix of scaled likelihoods.}

\item{KC}{a character vector of collections that the individuals belong to (or NA if you
don't know where they come from).  If this is NULL, then SL just gets returned untouched.}

\item{CFL}{the levels of the collections factor (which is within clean$short)}
}
\description{
This is used internally.
}
\keyword{internal}
