% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repuclust.R
\name{new_node}
\alias{new_node}
\title{given the top reclust mean pair, make a new node out of them}
\usage{
new_node(rmp, nodes, idx)
}
\arguments{
\item{rmp}{the output of top_repclust_mean_pair}

\item{nodes}{a list of nodes that are already there (so the members can be obtained)}

\item{idx}{the index that this node should take}
}
\description{
given the top reclust mean pair, make a new node out of them
}
\keyword{internal}
