% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_locus_means_and_vars.R
\name{per_locus_means_and_vars}
\alias{per_locus_means_and_vars}
\title{Compute the mean and variance of the single-locus genotype likelihoods for each collection}
\usage{
per_locus_means_and_vars(par_list)
}
\arguments{
\item{par_list}{genetic data converted to the param_list format by \code{tcf2param_list}. This should be
include genotypes only for the reference individuals.}
}
\value{
Returns a list with two components, mean and var, each one a matrix that has
C (number of collections) rows and L (number of loci) columns, giving the mean (or variance)
of the genotype likelihoods in the individuals in that collection at that locus.
}
\description{
This assumes that you have compiled params for a reference data set and then it just
calls rcpp_per_locus and then summarizes the results.
}
\keyword{internal}
