% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{perfect_chinook}
\alias{perfect_chinook}
\title{perfect-assignment genetic data for chinook.}
\source{
Made it up!
}
\description{
This is just like the \code{\link{chinook}} data, but only has 7 loci and all loci are
fixed in fortuitous patterns so that every single collection is easily resolved.  This is
primarily useful for testing purposes.
}
