% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{perfect_chinook_mix}
\alias{perfect_chinook_mix}
\title{perfect-assignment mixture genetic data for chinook.}
\source{
Made it up!
}
\description{
This is similar to the \code{\link{chinook_mix}} data, but only has 7 loci and all loci are
fixed in fortuitous patterns so that every single collection is easily resolved.  This is
primarily useful for testing purposes.  The name of the individual has its collection
inside the colons.
}
