% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_at_least_one_in_sample.R
\name{prob_at_least_one_in_sample}
\alias{prob_at_least_one_in_sample}
\title{Calculates the posterior probability that the sample contains at least one fish from each collection and repunit}
\usage{
prob_at_least_one_in_sample(X, mix_coll, burn_in = 100)
}
\arguments{
\item{X}{The return object from infer_mixture}

\item{mix_coll}{character name of the mixture_collection within X
you want to summarize the results for.}

\item{burn_in}{discards the first burn_in sweeps when handling the pi_traces
(for the boxplots of the pi)}
}
\description{
This takes the output of infer_mixture() and returns two tibbles, one for collections
and the other for repunits. Note that for this to work properly, infer_mixture() must
have been run with a total_catch_tib, and also it is imperative that the
variable_prob_is_catch parameter to infer_mixture() was at its default value of
FALSE.
}
