% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_close_matchers}
\alias{rcpp_close_matchers}
\title{Find all pairs that have close matches}
\usage{
rcpp_close_matchers(par_list, non_miss_fract, match_fract)
}
\arguments{
\item{par_list}{genetic data converted to the param_list format by \code{tcf2param_list}}
}
\value{
Gotta say more
}
\description{
Find all pairs that have close matches
}
\examples{
# gotta do something here too
}
\keyword{internal}
