% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_indiv_specific_logl_means_and_vars}
\alias{rcpp_indiv_specific_logl_means_and_vars}
\title{From the pattern of missing data at each individual, compute the expected mean and variance of the logl}
\usage{
rcpp_indiv_specific_logl_means_and_vars(par_list, MV)
}
\arguments{
\item{par_list}{genetic data converted to the param_list format by \code{tcf2param_list}}

\item{MV}{a list of two matrices, one of means and the other of variances, which are C x L
matrices.  This is basically the list that is returned by \code{\link{per_locus_means_and_vars}}.}
}
\value{
a matrix with C rows and I columns. Each row
represents a collection, and each column an individual.
}
\description{
This takes a param_list so that it has access to individual genotypes (and hence can cycle through them
and know which are missing and which are not.)  It also takes a matrix of per-locus logl means
and variances like what is computed by \code{\link{per_locus_means_and_vars}}.
}
\details{
This function doesn't do any checking to assure that the par_list and the per-locus logl means
matrix are made for one another.  (i.e. use the same collections in the same order.)
}
\keyword{internal}
