% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_per_locus_logls}
\alias{rcpp_per_locus_logls}
\title{Return a matrix of locus-specific self-assignment logls}
\usage{
rcpp_per_locus_logls(par_list)
}
\arguments{
\item{par_list}{genetic data converted to the param_list format by \code{tcf2param_list}}
}
\value{
\code{rcpp_per_locus_logls} returns a matrix with I rows and L columns. Each row
represents an individual, and each column a locus. Note that missing data at a locus
returns a zero.  That should be changed to NA later.
}
\description{
Takes a list of key parameters from a genetic dataset, and calculates
the log-likelihood of each individual's single-locus genotype, given the allele counts
in the individual's collection.
}
\details{
This uses Leave-One-Out cross-validation is used to avoid bias in log-likelihood for an
individual's known collection of origin
}
\keyword{internal}
