% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gsi_sim.R
\name{read_gsi_sim}
\alias{read_gsi_sim}
\title{read a gsi_sim formatted input file into a tibble that rubias can use}
\usage{
read_gsi_sim(path, sample_type, repunits = NULL)
}
\arguments{
\item{path}{path to the gsi_sim file}

\item{sample_type}{should be "reference" or "mixture" depending on what kind of file it is}

\item{repunits}{the gsi_sim reporting units file.  Has no effect if sample_type is "mixture".  If
sample_type is "reference" and this is left as NULL, then all collections will be put in the the
"default_repu".}
}
\description{
Note that this relies on a system call to awk.  It probably won't work
on Windows.
}
