% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_pipelines.R
\name{ref_and_mix_pipeline}
\alias{ref_and_mix_pipeline}
\title{Estimate mixing proportions from reference and mixture datasets}
\usage{
ref_and_mix_pipeline(reference, mixture, gen_start_col, method = "MCMC")
}
\arguments{
\item{reference}{a dataframe of two-column genetic format data, proceeded by "repunit", "collection",
and "indiv" columns. Does not need "sample_type" column, and will be overwritten if provided}

\item{mixture}{a dataframe of two-column genetic format data. Must have the same structure as
\code{reference} dataframe, but "collection" and "repunit" columns are ignored.
Does not need "sample_type" column, and will be overwritten if provided}

\item{gen_start_col}{the first column of genetic data in both data frames}

\item{method}{a choice between "MCMC", "PB" and "BH" methods for estimating mixture proportions}
}
\value{
\code{mix_proportion_pipeline} returns the standard output of the chosen
mixing proportion estimation method (always a list). For method "PB",
returns the standard MCMC results, as well as the bootstrap-corrected
means under \code{$mean$bootstrap}
}
\description{
Takes a mixture and reference dataframe of two-column genetic data, and a
desired method of estimation for the population mixture proportions (MCMC, PB, or BH MCMC)
Returns the output of the chosen estimation method
}
\details{
"MCMC" estimates mixing proportions and individual posterior
probabilities of assignment through Markov-chain Monte Carlo,
while "PB" does the same with a parametric bootstrapping correction,
and "BH" uses the misassignment-scaled, hierarchial MCMC.
}
\examples{
reference <- alewife[,-1]
mixture <- alewife[,-1]
gen_start_col <- 14
bh <- ref_and_mix_pipeline(reference, mixture, gen_start_col, method = "BH")
mcmc <- ref_and_mix_pipeline(reference, mixture, gen_start_col, method = "MCMC")
PB <- ref_and_mix_pipeline(reference, mixture, gen_start_col, method = "PB")

}

