% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repuclust.R
\name{repuclust}
\alias{repuclust}
\title{perform greedy clustering into reporting units to visualize relationships}
\usage{
repuclust(reference, gen_start_col)
}
\description{
perform greedy clustering into reporting units to visualize relationships
}

