% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repuclust.R
\name{repuclust2phylo}
\alias{repuclust2phylo}
\title{convert the output of repuclust to an ape phylo object}
\usage{
repuclust2phylo(r)
}
\arguments{
\item{r}{the output from repuclust()}
}
\description{
convert the output of repuclust to an ape phylo object
}
