% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmultinom_1}
\alias{rmultinom_1}
\title{Simulate a single multinomial vector from within Rcpp}
\usage{
rmultinom_1(size, probs, N)
}
\arguments{
\item{size}{the number of trials}

\item{probs}{the cell probabilities}

\item{N}{the number of cells}
}
\value{
An IntegerVector of length N
}
\description{
From: https://gallery.rcpp.org/articles/recreating-rmultinom-and-rpois-with-rcpp/
}
\keyword{internal}
