% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{round2}
\alias{round2}
\title{Round a given number, with 5 always rounded up}
\usage{
round2(x, n)
}
\arguments{
\item{x}{the data to be rounded}

\item{n}{the number of digits to round to}
}
\description{
Given a number and a digit to round to, returns the rounded number,
with 5 always rounded upwards.
}
\details{
This function differs from \code{round}, which rounds 5 "towards the
even number". Rounding 5s up leads to bias when positive and negative numbers
are expected, but can be desired in some cases.
}
\keyword{internal}
