% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_spec_examples}
\alias{sim_spec_examples}
\title{List of example ways of specifying repunit and collection quantities in simulations}
\source{
Made it up!
}
\description{
This is just a list of tibbles that can be passed to the alpha_repunit or the
alpha_collection parameters in, for example, \code{\link{assess_reference_loo}}.
}
