% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_chinook_mix}
\alias{small_chinook_mix}
\title{Small sample of SNP data from Chinook salmon taken in May/August 2015 from California fisheries}
\source{
Southwest Fisheries Science Center, Santa Cruz, CA
}
\description{
This is simply a sample of 100 fish from \code{\link{chinook}}.
}
