% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tabulate_allocations}
\alias{tabulate_allocations}
\title{Given a vector of n different categories in 1...n, count up their occurrences and return in a vector of length n}
\usage{
tabulate_allocations(C, n)
}
\arguments{
\item{C}{a vector of categories taking values of 1,...,n}

\item{n}{the number of categories}
}
\description{
I should have written this for the other functions above, but I am just getting to it now
to sum up the Zeds of the fish for the total catch sampling stuff.
}
\keyword{internal}
