% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eca-internal-funcs.R
\name{tidy_mcmc_coll_rep_stuff}
\alias{tidy_mcmc_coll_rep_stuff}
\title{A helper function to tidy up the output from the gsi_mcmc functions}
\usage{
tidy_mcmc_coll_rep_stuff(field, p, pname, car_tib)
}
\arguments{
\item{field}{The output to tidy (i.e.. out$mean)}

\item{p}{the name of the parameter whose values you want to extract (like "pi")}

\item{pname}{the name that you want the parameter to be called in the output}

\item{car_tib}{a tibble with repunit and collection in the order they appear in the output}
}
\description{
This makes a tidy data frame of stuff, and also changes things back to
factors, if the levels are provided.
}
\keyword{internal}
