% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repuclust.R
\name{top_repclust_mean_pair}
\alias{top_repclust_mean_pair}
\title{find the pair that includes the highest incorrect normalized scaled likelihood}
\usage{
top_repclust_mean_pair(pp)
}
\arguments{
\item{pp}{a data frame like pair_posts}
}
\description{
find the pair that includes the highest incorrect normalized scaled likelihood
}
\keyword{internal}

