% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{turn_non_catch_to_minus_one}
\alias{turn_non_catch_to_minus_one}
\title{Simulate whether fish are part of the catch or not}
\usage{
turn_non_catch_to_minus_one(Z, P, NC)
}
\arguments{
\item{Z}{a vector of categories taking values of 1,...,n}

\item{P}{probabilities of being considered "catch" for each of the mixure samples}

\item{NC}{output by reference of the number of fish considered to actually be catch.}
}
\description{
This is for the variable_prob_is_catch == TRUE scenario. We simply
go through the vector Z of allocations. For each one, we simulate
a uniform RV.  If that RV is greater than the corresponding term
in P, then we turn that Z into a -1, so it will not be counted by
tabulate_allocations().  NC returns the number that are part of the
catch by reference.
}
\keyword{internal}
