% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_dp}
\alias{update_dp}
\title{Update a matrix of Dirichlet parameters given a set of collection allocations}
\usage{
update_dp(par_list, allocations)
}
\arguments{
\item{par_list}{genetic data converted to the param_list format by \code{tcf2param_list}}

\item{allocations}{an N-long vector of collection ID numbers from \code{samp_from_post}}
}
\value{
\code{geno_logL} returns a matrix with C rows and I columns. Each column
represents an individual, and each row the log-likelihood of that individual's
genotype, given the allele counts in that collection
}
\description{
Takes a list of key parameters from a genetic dataset, as well as a vector
of collections to which individuals have been allocated, and updates the baseline
allele frequencies as if the individuals were known to originate from said collections
}
\details{
For each individual and locus, this identifies the alleles held by that individual,
then adds 1 to these alleles' counts in the Dirichlet parameter vector
}
\keyword{internal}
