% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_reference_loo.R
\name{assess_reference_loo}
\alias{assess_reference_loo}
\title{Simulate mixtures and estimate reporting group and collection proportions.}
\usage{
assess_reference_loo(reference, gen_start_col, reps = 50, mixsize = 100,
  seed = 5, alpha_repunit = 1.5, alpha_collection = 1.5,
  resampling_unit = "individual", printSummary = FALSE)
}
\arguments{
\item{reference}{a two-column format genetic dataset, with "repunit", "collection", and "indiv"
columns, as well as a "sample_type" column that has some "reference" entries}

\item{gen_start_col}{the first column of genetic data in \code{reference}}

\item{reps}{number of reps of mixture simulation and MCMC to do}

\item{mixsize}{the number of individuals in each simulated mixture}

\item{seed}{a random seed for simulations}

\item{alpha_repunit}{If a vector, this is the dirichlet parameter for simulating
the proportions of reporting units. Gets recycled to the number of reporting units. Default is 1.5.
Otherwise, this could be a two-column data frame.  The first column must be named "repunit" and the
second one must be one of "dirichlet", "ppn", or "cnt", according to whether you wish to
specify dirichlet parameters, or proportions, or exact counts, respectively, for each population.
If you want to make multiple simulations, pass in a list of data frames or of individual dirichlet parameters.}

\item{alpha_collection}{The dirichlet parameter for simulating proportions of collections within reporting units. Default = 1.5.
If this is a data frame then the first column must be "collection" and the second must be one of
"dirichlet", "ppn", "cnt", "sub_dirichlet", "sub_ppn".  If you want to provide multiple different
scenarios.  You can pass them in as a list.  If alpha_repunit or alpha_collection is a list with length
greater than 1, the shorter will be recycled.}

\item{resampling_unit}{what unit should be resampled.  Currently the choices are "individuals" (the default)
and "gene_copies".  Using "individuals" preserves missing data patterns available in the reference data set.
We also have "gene_copies_with_missing" capability, but it is not yet linked into this function.}

\item{printSummary}{if TRUE a summary of the reference samples will be printed to stdout.}
}
\description{
From a reference dataset, this creates a genotype-logL matrix based on
simulation-by-individual with randomly drawn population proportions,
then uses this in two different estimates of population mixture proportions:
maximum likelihood via EM-algorithm and posterior mean from
MCMC.
}
\details{
We still need to implement storage of individual-specific results for the simulated individuals.
This we can do pretty easily, but wanted to get the proportions done and out there first.  Send eric some
email if you really need individual specific output and he will prioritize it.
}
\examples{
ale_dev <- assess_reference_loo(alewife, 17)

}
