% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eca-funcs.R
\name{assess_reference_mc}
\alias{assess_reference_mc}
\title{Simulate mixtures and estimate reporting group and collection proportion estimation.}
\usage{
assess_reference_mc(reference, gen_start_col, reps = 50, mixsize = 100,
  seed = 5, alpha_repunit = 1.5, alpha_collection = 1.5,
  min_remaining = 5)
}
\arguments{
\item{reference}{a two-column format genetic dataset, with "repunit", "collection", and "indiv"
columns, as well as a "sample_type" column that has some "reference" entries.}

\item{gen_start_col}{the first column of genetic data in reference}

\item{reps}{number of reps to do}

\item{mixsize}{the number of individuals in each simulated mixture.}

\item{seed}{a random seed for simulations}

\item{alpha_repunit}{The dirichlet parameter for simulating the proportions of reporting units. Default = 1.5}

\item{alpha_collection}{The dirichlet parameter for simulating proportions of collections within reporting units. Default = 1.5}
}
\description{
From a reference dataset, this creates a genotype-logL matrix based on
simulation-by-individual with randomly drawn population proportions,
then uses this in two different estimates of population mixture proportions:
maximum likelihood via EM-algorithm and posterior mean from
MCMC.
}
\details{
This is hard-wired at the moment to do something like Hasselman et al.
}
\examples{
ale_dev <- assess_reference_loo(alewife, 15)

}

