% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_correction.R, R/simulation_pipeline.R
\name{bias_comparison}
\alias{bias_comparison}
\title{Test the effects of bias correction on a particular reference dataset}
\usage{
bias_comparison(reference, gen_start_col, seed = 10)

bias_comparison(reference, gen_start_col, seed = 10)
}
\arguments{
\item{reference}{a two-column format genetic dataset, with a "repunit" column
specifying each individual's reporting unit of origin, a "collection" column
specifying the collection (population or time of sampling) and "indiv" providing
a unique name}

\item{gen_start_col}{the first column containing genetic data in \code{reference}.
All columns should be genetic format following this column, and gene copies from the
same locus should be adjacent}

\item{seed}{the random seed for simulations}

\item{reference}{a two-column format genetic dataset, with a "repunit" column
specifying each individual's reporting unit of origin, a "collection" column
specifying the collection (population or time of sampling) and "indiv" providing
a unique name}

\item{gen_start_col}{the first column containing genetic data in \code{reference}.
All columns should be genetic format following this column, and gene copies from the
same locus should be adjacent}

\item{seed}{the random seed for simulations}
}
\value{
\code{bias_comparison} returns a list; the first element is
a list of the relevant rho values generated on each iteration of the random "mixture"
creation. This includes the true rho value, the standard result \code{rho_mcmc},
the misassignment-scaled \code{rho_bh}, and the parametric bootstrapped \code{rho_pb}.

The second element is a dataframe listing summary statistics for each
reporting unit and estimation method. \code{mse}, the mean squared error, summarizes
the deviation of the rho estimates from their true value, including both bias and other variance.
\code{mean_prop_bias} is the average ratio of residual to true value, which gives greater
weight to deviations at smaller values. \code{mean_bias} is simply the average residual;
unlike \code{mse}, this demonstrates the direction of the bias.

\code{bias_comparison} returns a list; the first element is
a list of the relevant rho values generated on each iteration of the random "mixture"
creation. This includes the true rho value, the standard result \code{rho_mcmc},
the misassignment-scaled \code{rho_bh}, and the parametric bootstrapped \code{rho_pb}.

The second element is a dataframe listing summary statistics for each
reporting unit and estimation method. \code{mse}, the mean squared error, summarizes
the deviation of the rho estimates from their true value, including both bias and other variance.
\code{mean_prop_bias} is the average ratio of residual to true value, which gives greater
weight to deviations at smaller values. \code{mean_bias} is simply the average residual;
unlike \code{mse}, this demonstrates the direction of the bias.
}
\description{
Takes a reference two-column genetic dataset, pulls a series of random
"mixture" datasets with varying reporting unit proportions from this reference,
and compares the results of GSI through standard MCMC, misassignment-scaled MCMC,
and parametric-bootstrap MCMC bias correction

Takes a reference two-column genetic dataset, pulls a series of random
"mixture" datasets with varying reporting unit proportions from this reference,
and compares the results of GSI through standard MCMC, misassignment-scaled MCMC,
and parametric-bootstrap MCMC bias correction
}
\details{
The amount of bias in reporting unit proportion calculations increases with the
rate of missassignment between reporting units (decreases with genetic differentiation),
and increases as the number of collections within reporting units becomes more uneven.

Output from the standard Bayesian MCMC method demonstrates the level of bias to be
expected for the input dataset; parametric bootstrapping is an empirical method
for the removal of any existing bias, while the misassignment-scaled MCMC is a semi-empirical
method based on the rate of misassignment, which takes into account both genetic differentiation
and uneven collection distribution.

The amount of bias in reporting unit proportion calculations increases with the
rate of missassignment between reporting units (decreases with genetic differentiation),
and increases as the number of collections within reporting units becomes more uneven.

Output from the standard Bayesian MCMC method demonstrates the level of bias to be
expected for the input dataset; parametric bootstrapping is an empirical method
for the removal of any existing bias, while the misassignment-scaled MCMC is a semi-empirical
method based on the rate of misassignment, which takes into account both genetic differentiation
and uneven collection distribution.
}
\examples{
ale_bias <- bias_comparison(alewife, 15)

ale_bias <- bias_comparison(alewife, 15)

}

