% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blueback}
\alias{blueback}
\title{Microsat data from blueback herring reference populations.}
\format{A tbl_df-ed (from dplyr) data frame with XXX rows and YY variables. The relevant variables for
analyses here are:
\describe{
  \item{Field_Num}{unique identifier for each bird}
  \item{Isotope.Value}{hydrogen isotope ratios measured in the bird's feather}
  \item{lat}{latitude of the bird's breeding/sampling location}
  \item{long}{latitude of the bird's breeding/sampling location}
}}
\source{
Got it off GitHub, but must be at Dryad now too...
}
\description{
gotta say more about it
}
