% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_reference_z_scores.R
\name{compute_reference_z_scores}
\alias{compute_reference_z_scores}
\title{compute empirical Z-scores for reference population individuals}
\usage{
compute_reference_z_scores(SA)
}
\arguments{
\item{SA}{The tibble that is output by \code{\link{self_assign}}.}
}
\value{
A tibble with a few more columns
}
\description{
For each collection, we compute the mean and sd of the log-likelihoods
of assigning correctly to the collection.  We then compute the Z-values
for each individual.
}
