% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{geno_logL}
\alias{geno_logL}
\title{Calculate a matrix of genotype log-likelihoods for a genetic dataset}
\usage{
geno_logL(par_list)
}
\arguments{
\item{par_list}{genetic data converted to the param_list format by \code{tcf2param_list}}
}
\value{
\code{geno_logL} returns a matrix with C rows and I columns. Each column
represents an individual, and each row the log-likelihood of that individual's
genotype, given the allele counts in that collection
}
\description{
Takes a list of key parameters from a genetic dataset, and calculates
the log-likelihood of each individual's genotype, given the allele counts
in each collection
}
\details{
Leave-One-Out cross-validation is used to avoid bias in log-likelihood for an
individual's known collection of origin
}
\examples{
example(tcf2param_list)
ale_glL <- geno_logL(ale_par_list)
}
\keyword{internal}
