% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gprob_sim_gc_missing}
\alias{gprob_sim_gc_missing}
\title{Simulate genotypes by gene copy, with missing data from chosen individuals}
\usage{
gprob_sim_gc_missing(par_list, sim_colls, sim_missing)
}
\arguments{
\item{par_list}{list_diploid_params output}

\item{sim_colls}{a vector length I of collections from which to sample the genotypes
for individual i}

\item{sim_missing}{a vector of length I of indices for individuals in I_list
whose missing data should be copied for individual i}
}
\description{
Simulate genotypes by gene copy, with missing data from chosen individuals
}

