% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gsi_mcmc_1}
\alias{gsi_mcmc_1}
\title{do MCMC from the simplest GSI model for pi and the individual posterior probabilities}
\usage{
gsi_mcmc_1(SL, Pi_init, lambda, reps, burn_in, sample_int_Pi, sample_int_PofZ)
}
\arguments{
\item{SL}{matrix of the scaled likelihoods.  This is should have values for each individual in a column
(going down in the rows are values for different populations).}

\item{Pi_init}{Starting value for the pi vector.}

\item{reps}{total number of reps (sweeps) to do.}

\item{burn_in}{how many reps to discard in the beginning when doing the mean calculation. They will still be
returned in the traces if desired}

\item{sample_int_Pi}{the number of reps between samples being taken for Pi traces.  If 0 no trace samples are taken}

\item{sample_int_PofZ}{the number of reps between samples being taken for the traces of posterior of each individual's origin. If 0
no trace samples are taken.}
}
\description{
Using a matrix of scaled likelihoods this function samples values of pi and the posteriors
for all the individuals.  It returns the output in a list.
}
\examples{
params <- tcf2param_list(alewife, 15)
logl <- geno_logL(params)
SL <- apply(exp(logl), 2, function(x) x/sum(x))
lambda <- rep(1/params$C, params$C)
mcmc <- test_gsi_mcmc_1(SL, lambda, lambda, 10000, 2500, 50, 50)

}

