% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_pipeline.R
\name{Hasselman_sim_colls}
\alias{Hasselman_sim_colls}
\title{Generate a random sim_colls vector for simulation of individual genotypes, based on the methods
used in Hasselman et al. 2015, and the collection-repunit relationship vectors generated from a
reference}
\usage{
Hasselman_sim_colls(RU_starts, RU_vec)
}
\description{
Generate a random sim_colls vector for simulation of individual genotypes, based on the methods
used in Hasselman et al. 2015, and the collection-repunit relationship vectors generated from a
reference
}

