% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_pipelines.R
\name{Hasselman_simulation_pipeline}
\alias{Hasselman_simulation_pipeline}
\title{Recreate the simulations of Hasselman \emph{et al.} (2015) with two alternative
GSI methods}
\usage{
Hasselman_simulation_pipeline(reference, gen_start_col, seed = 5)
}
\arguments{
\item{reference}{a two-column format genetic dataset, with "repunit", "collection", and "indiv"
columns}

\item{gen_start_col}{the first column of genetic data in reference}

\item{seed}{a random seed for simulations}
}
\description{
Using a reference dataset, creates a genotype-logL matrix based on
simulation-by-individual with randomly drawn population proportions,
then uses this in three different estimates of population mixture proportions:
MCMC, MCMC corrected with parametric bootstrapping, and hierarchical MCMC
}
\details{
The reference data set is processed with \code{tcf2param_list},
and the average-correct-assignment matrix is calculated using
\code{avg_coll2correctRU}.

50 Hasselman-style simulated mixture samples are then created, run through MCMC, PB,
and hierarchical MCMC GSI, and plotted against the rho value used to simulate
the mixture.
}
\examples{
ale_dev <- Hasselman_simulation_pipeline(alewife, 15)

}

