% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversion.R
\name{list_diploid_params}
\alias{list_diploid_params}
\title{Collect essential data values before mixture proportion estimation}
\usage{
list_diploid_params(AC_list, I_list, PO, coll_N, RU_vec, RU_starts)
}
\arguments{
\item{AC_list}{output from reference_allele_counts}

\item{I_list}{output from allelic_list}

\item{PO}{vector of collection (population of origin) indices
for every individual in the sample}

\item{coll_N}{a vector of the total number of individuals in each collection}

\item{RU_vec}{a vector of collection indexes, sorted by reporting unit}

\item{RU_starts}{a vector of indexes, designating the first collection for each
reporting unit in RU_vec}
}
\value{
\code{list_diploid_params} returns a list of the information necessary
for the calculation of genotype likelihoods in MCMC
}
\description{
Collect essential data values before mixture proportion estimation
}
\examples{
example(allelic_list)
PO <- as.integer(factor(ale_long$clean_short$collection))
coll_N <- dplyr::count(ale_long$clean_short, collection) \%>\%
  dplyr::select(n)
  coll_N <- unname(unlist(coll_N))
Colls_by_RU <- dplyr::count(ale_long$clean_short, repunit, collection) \%>\%
   dplyr::select(-n)
 PC <- rep(0, length(unique((Colls_by_RU$repunit))))
 for(i in 1:nrow(Colls_by_RU)) {
   PC[Colls_by_RU$repunit[i]] <- PC[Colls_by_RU$repunit[i]] + 1
 }
RU_starts <- c(0, cumsum(PC))
RU_vec <- as.integer(Colls_by_RU$collection)
param_list <- list_diploid_params(ale_ac, ale_alle_list, PO, coll_N, RU_vec, RU_starts)

}

