% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mixture_draw}
\alias{mixture_draw}
\title{Separate a chosen proportion of a reference dataset into a mixture with known population proportions}
\usage{
mixture_draw(D, rhos = NULL, omegas = NULL, N,
  min_remaining = 1/(length(omegas) * 10))
}
\arguments{
\item{D}{a two-column genetic dataframe with "indiv", "repunit", and "collection" columns}

\item{rhos}{a vector of the desired reporting unit proportions in the mixture set;
if not named, will be assumed to be ordered by order of appearance in the dataset}

\item{omegas}{the desired collection proportions in the mixture set}

\item{N}{the total size of the mixture set}

\item{min_remaining}{the fraction of any collection in the reference dataset which must remain
at the end of the draw}
}
\value{
a list of two data frames, "mixture" being the random sample taken, and "reference" being the remaining samples

D <- alewife
N <- 100
rhos <- as.vector(gtools::rdirichlet(1, table(alewife$repunit)))
min_remaining <- .005
}
\description{
Separate a chosen proportion of a reference dataset into a mixture with known population proportions
}

