% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_pipelines.R
\name{ref_and_mix_pipeline}
\alias{ref_and_mix_pipeline}
\title{Estimate mixing proportions from reference and mixture datasets}
\usage{
ref_and_mix_pipeline(reference, mixture, gen_start_col, method = "MCMC")
}
\arguments{
\item{reference}{a dataframe of two-column genetic format data, proceeded by "repunit", "collection",
and "indiv" columns. Does not need "sample_type" column, and will be overwritten if provided}

\item{mixture}{a dataframe of two-column genetic format data. Must have the same structure as
\code{reference} dataframe, but "collection" and "repunit" columns are ignored.
Does not need "sample_type" column, and will be overwritten if provided}

\item{gen_start_col}{the first column of genetic data in both data frames}

\item{method}{a choice between "EM", "MCMC", and "BH" methods for estimating mixture proportions}
}
\value{
\code{mix_proportion_pipeline} returns the standard output of the chosen
mixing proportion estimation method (always a list)
}
\description{
Takes a mixture and reference dataframe of two-column genetic data, and a
desired method of estimation for the population mixture proportions (EM, MCMC, or BH MCMC)
Returns the output of the chosen estimation method
}
\details{
"EM" estimates mixing proportions and individual posterior
probabilities of assignment through a simple expectation maximization,
while "MCMC" does the same with Markov-chain Monte Carlo, and "BH" uses the misassignment-scaled,
hierarchial MCMC.
}
\examples{
reference <- alewife[,-1]
mixture <- alewife[,-1]
gen_start_col <- 14
bh <- mix_proportion_pipeline(reference, mixture, gen_start_col, method = "BH")
mcmc <- mix_proportion_pipeline(reference, mixture, gen_start_col, method = "MCMC")
em <- mix_proportion_pipeline(reference, mixture, gen_start_col, method = "EM")

}

