% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repuclust.R
\name{repuclust}
\alias{repuclust}
\title{perform greedy clustering into reporting units to visualize relationships}
\usage{
repuclust(reference, gen_start_col)
}
\arguments{
\item{reference}{a two-column format genetic dataset, with "repunit", "collection", and "indiv"
columns, as well as a "sample_type" column that has some "reference" entries}

\item{gen_start_col}{the first column of genetic data in \code{reference}}
}
\description{
perform greedy clustering into reporting units to visualize relationships
}
