% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{round2}
\alias{round2}
\title{Rounding 5 up, rather than rounding to the nearest even number}
\usage{
round2(x, n)
}
\arguments{
\item{x}{the data to be rounded}

\item{n}{the number of digits to round to}
}
\description{
Sometimes desired, especially when the numbers are all positive,
and not possible with R (stolen from the Statistically Significant blog)
}

