% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubias.R
\docType{package}
\name{rubias}
\alias{rubias}
\alias{rubias-package}
\title{rubias: A package for bias correction in hierarchical GSI}
\description{
Blah blah blah.
}
\section{rubias functions}{

The rubias functions ...
}

