% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubias.R
\docType{package}
\name{rubias}
\alias{rubias}
\alias{rubias-package}
\title{rubias: A package for bias correction in hierarchical GSI}
\description{
Blah blah blah.
}
\section{rubias functions}{

The rubias functions ...
}

\section{example data}{

\code{alewife} & \code{blueback} are examples of good times to use rubias; there are
clearly observable biases based on the number of collections in a reporting unit.
 \code{chinook} is an example of a dataset in which examination of the graphs from
\code{bias_comparison} reveal that rubias is not appropriate; there was no appreciable
bias to begin with, and so both the misassignment-scaling and parametric boostrap methods
have no consistent effect on the mean residual, and often increase mean squared error
}

