% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{samp_from_mat}
\alias{samp_from_mat}
\title{Sample 1 observation from cell probabilities that are columns of a matrix}
\usage{
samp_from_mat(M)
}
\arguments{
\item{M}{a matrix whose columns are reals summing to one}
}
\description{
blah blah...
}

