% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_assign.R
\name{self_assign}
\alias{self_assign}
\title{Do leave-one-out self-assignment of individuals in a reference baseline}
\usage{
self_assign(reference, gen_start_col)
}
\arguments{
\item{reference}{a two-column format genetic dataset, with "repunit", "collection", and "indiv"
columns, as well as a "sample_type" column that has some "reference" entries}

\item{gen_start_col}{the first column of genetic data in \code{reference}}
}
\value{
a tibble ...
}
\description{
Returns a tidy data frame
}
\examples{
ale_sa <- self_assign(alewife, 17)
}
