% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eca-internal-funcs.R
\name{simulate_random_samples}
\alias{simulate_random_samples}
\title{Generate a samples for a mixture.}
\usage{
simulate_random_samples(RU_starts, RU_vec, size = 100, alpha_repunit = 1.5,
  alpha_collection = 1.5)
}
\arguments{
\item{RU_starts}{a vector delineating the reporting units in \code{RU_vec};
generated by \code{tcf2param_list}}

\item{RU_vec}{a vector of collection indices, grouped by reporting unit;
generated by \code{tcf2param_list}}

\item{size}{The number of individuals desired in the mixture sample.  Default = 100.}

\item{alpha_repunit}{The dirichlet parameter for simulating the proportions of reporting units. Default = 1.5}

\item{alpha_collection}{The dirichlet parameter for simulating proportions of collections within reporting units. Default = 1.5}
}
\value{
a list with three elements.
The first two are a rho vector and an omega vector, respectively. The third is a vector of origins for
simulated individuals, sampled from the collections with probabilities = omega
}
\description{
Creates random reporting unit (rho) and collection (omega) proportions, and a
\code{sim_colls} vector for simulation of individual genotypes, based on the methods
used in Hasselman \emph{et al.} (2015)
}
\keyword{internal}

