% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversion.R
\name{tcf2param_list}
\alias{tcf2param_list}
\title{Generate MCMC parameter list from two-column genetic data & Print Summary}
\usage{
tcf2param_list(D, gen_start_col, samp_type = "both", summ = T)
}
\arguments{
\item{D}{A data frame containing two-column genetic data, preceded by metadata.
The header of the first genetic data column in each pair lists the locus name,
the second is ignored. \strong{Locus names must not have spaces in them!}
Required metadata includes a column of unique individual identifiers named "indiv",
a column named "collection" designating the sample groups, and a "sample_type" column
denoting each individual as a "reference" or "mixture" sample}

\item{gen_start_col}{The index (number) of the column in which genetic data starts.
Columns must be only genetic data after genetic data starts.}

\item{samp_type}{the sample groups to be include in the individual genotype list,
whose likelihoods will be used in MCMC. Options "reference", "mixture", and "both"}

\item{summ}{(LOGICAL) should summary descriptions of the formatted data be provided?}
}
\description{
This function is a wrapper for all steps to create the parameter list necessary
for genotype log-likelihood calculation from the starting two-column genetic data
}
\examples{
ale_par_list <- tcf2param_list(alewife, 15)

}

