% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eca-internal-funcs.R
\name{tidy_pi_traces}
\alias{tidy_pi_traces}
\title{a helper function to tidy up the pi-traces that come out of the mcmc functions}
\usage{
tidy_pi_traces(input, pname, car_tib, coll_levs, repu_levs, interval)
}
\arguments{
\item{input}{The output to tidy (i.e.. out$trace$pi)}

\item{pname}{the name that you want the parameter to be called in the output}

\item{car_tib}{a tibble with repunit and collection in the order they appear in the output}

\item{coll_levs}{a vector of levels for the collection (or NULL if none)}

\item{repu_levs}{a vector of levels for the repunit (or NULL if none)}

\item{interval}{the thinning interval that was used}

\item{mix_indiv_tib}{a tibble with the individuals in the order they appear in the output}
}
\description{
This makes a tidy data frame of stuff, and also changes things back to
factors, if the levels are provided.
}
\keyword{internal}
