% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gsi_sim_mixture.R
\name{write_gsi_sim_mixture}
\alias{write_gsi_sim_mixture}
\title{Write a mixerence data frame to gsi_sim format baseline and repunits file}
\usage{
write_gsi_sim_mixture(mix, gen_start_col, mixprefix = "mixture")
}
\arguments{
\item{mix}{mixture data frame}

\item{gen_start_col}{column in which the genetic data start}

\item{mixprefix}{path to write the mixture file to.  The mixture collection name + .txt will
be appended to this.}
}
\description{
Note, this is only intended to work with integer-valued alleles, at the moment.
It was just written for testing and verifying that things are working correctly.
}
