% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gsi_sim_reference.R
\name{write_gsi_sim_reference}
\alias{write_gsi_sim_reference}
\title{Write a reference data frame to gsi_sim format baseline and repunits file}
\usage{
write_gsi_sim_reference(ref, gen_start_col, baseout = "baseline.txt",
  repout = "repunits.txt")
}
\arguments{
\item{ref}{reference data frame}

\item{gen_start_col}{column in which the genetic data start}

\item{baseout}{path to write the baseline file to}

\item{repout}{path to write the repunits file to}
}
\description{
Note, this is only intended to work with integer-valued alleles, at the moment.
It was just written for testing and verifying that things are working correctly.
}
